The graphical environment has been improoved in order to show cells and obstacles dinamically. Some state variables like mapSize and the  obstacles coordinates has been added to the message sended via libwebsocket to the environment.

+++ Environment_DSE/fmu-pvs/misraC/Environment_old.c
@@ -59,8 +59,8 @@ void init(State* st) {
 		
 		x = (float64_t*)malloc(st->nRobots*sizeof(float64_t));
 		y = (float64_t*)malloc(st->nRobots*sizeof(float64_t));
-		ox = (float64_t*)malloc(st->nObstacles*sizeof(int32_t));
-		oy = (float64_t*)malloc(st->nObstacles*sizeof(int32_t));
+		ox = (int32_t*)malloc(st->nObstacles*sizeof(int32_t));
+		oy = (int32_t*)malloc(st->nObstacles*sizeof(int32_t));
 		
 		epslon = unifRand();
 		isInit = TRUE;
